+++ lib/audio_manager.dart
@@ -2,4 +2,4 @@ library audio_manager;
 
 export './src/index.dart';
 export './src/AudioInfo.dart';
-export './src/PlayMode.dart';
+export './src/AudioType.dart';