+++ lib/src/PlayMode.dart
@@ -1,5 +1,0 @@
-enum PlayMode {
-  sequence,
-  shuffle,
-  single,
-}